// heroes-app-main/src/app/heroes/interfaces/hero.interface.ts
// Generated by https://quicktype.io

import { SafeResourceUrl } from "@angular/platform-browser";

export interface Hero {
  id:               string;
  superhero:        string;
  publisher:        Publisher;
  alter_ego:        string;
  first_appearance: string;
  characters:       string;
  alt_img?:         string;
  upload_img?:      SafeResourceUrl;
}

export enum Publisher {
  DCComics = "DC Comics",
  MarvelComics = "Marvel Comics",
}